For #23424 - Part 1: Refactor TopSiteItemMenu into its own file
